fix: move plugin-sync to top level
You can still put related functionality in a folder with the same name
